full_stack_web_developer/05_capstone/requirements.txt
alembic==1.4.2
astroid==2.2.5
Click==7.0
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
Flask-Script==2.0.6
future==0.17.1
gunicorn==20.0.4
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.7
pycodestyle==2.5.0
pycosat==0.6.3
pycryptodome==3.3.1
python-dotenv==0.13.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
urllib3==1.25.6
Werkzeug==0.15.2
wrapt==1.11.1
